pyproject.toml
[project]
name = "email-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "fastembed>=0.7.4",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "google-cloud-discoveryengine>=0.15.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-speech>=2.34.0",
    "google-genai>=1.53.0",
    "jinja2>=3.1.6",
    "langchain>=1.1.2",
    "langchain-google-community>=3.0.2",
    "langchain-google-genai>=3.2.0",
    "langchain-google-vertexai>=3.1.1",
    "langchain-huggingface>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "langsmith>=0.4.56",
    "pillow>=11.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf2>=3.0.1",
    "unstructured>=0.18.21",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "mypy>=1.19.0",
    "ruff>=0.14.8",
]
